BIQUB1_Marlin2.0.7+H2+TitaniumHighTemp+BLTOUCH
Full code edit for the Biqu B1 with the added H2 extruder with a titanium heatbreak for up to 300 (285) degrees max limit and BL Touch for homing and bed leveling. 
Edge probing offset is 50mm because of the mount im currently using. Will update new firmware when add the new, custom designed (by me) mount for the hotend and BL Touch. Mounts and such will be shared on Thingverse and my own site when I have completed testing them.